#2308 extract lazy array access into a trait
